chore: Clean up warnings from linters in main.cpp.

+++ include/config.h
@@ -3,8 +3,6 @@
 #include "driver/i2c.h"
 #include <MD_Parola.h>
 #include <chrono>
-#include <cstdint>
-#include <string_view>
 
 // GPIO pin for built-in LED
 constexpr int CONFIG_BLINK_GPIO = 2;
@@ -17,8 +15,8 @@ constexpr int NTP_UPDATE_INTERVAL_MS = 60 * 60 * 1000; // 1 hour
 constexpr char TIMEZONE[] = "CET-1CEST,M3.5.0/2,M10.5.0/3";
 
 // Intervals and durations
-constexpr int STATUS_UPDATE_INTERVAL_SECONDS{5};
-constexpr int FORECAST_MINIMAL_DISPLAY_TIME_SECONDS{3};
+constexpr uint8_t STATUS_UPDATE_INTERVAL_SECONDS{5};
+constexpr uint8_t FORECAST_MINIMAL_DISPLAY_TIME_SECONDS{3};
 constexpr uint8_t FORECAST_PAGE_SWITCH_PROXIMITY{12}; // proximity threshold to switch forecast pages
 constexpr uint8_t PRECIPITATION_PAGE_SWITCH_PROXIMITY{100}; // proximity threshold for precipitation chart
 constexpr char FORECAST_API_URL[] =
